restructure class follow mvvm pattern
