+++ app/src/main/java/com/singtel/groupit/view/activity/LoginActivity.java
@@ -1,10 +1,9 @@
-package com.singtel.groupit;
+package com.singtel.groupit.view.activity;
 
 import android.content.Context;
-import android.content.Intent;
 import android.os.Bundle;
-import android.os.Handler;
-import android.os.Message;
+
+import com.singtel.groupit.R;
 
 import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper;
