+++ app/src/main/res/layout/activity_login.xml
@@ -11,7 +11,7 @@
     android:paddingEnd="@dimen/activity_horizontal_margin"
     android:paddingStart="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_horizontal_margin"
-    tools:context="com.singtel.groupit.MainActivity"
+    tools:context="com.singtel.groupit.view.activity.MainActivity"
     tools:ignore="MissingPrefix">
 
     <ImageView
