+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:id="@+id/activity_main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.singtel.groupit.MainActivity">
+    tools:context="com.singtel.groupit.view.activity.MainActivity">
 
     <TextView
         android:id="@+id/tv_test"
