+++ app/src/main/res/layout/activity_splash.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context="com.singtel.groupit.MainActivity">
+    tools:context="com.singtel.groupit.view.activity.MainActivity">
 
 
     <ImageView
